[1647] Cleanup for tests & mixins for the UAPI Django application
Part 1: Move mixins to the appropriate file
